git subrepo clone --force https://github.com/jonathanstowe/p6-Linux-Fuser.git jonathanstowe/p6-Linux-Fuser
subrepo:
  subdir:   "jonathanstowe/p6-Linux-Fuser"
  merged:   "98b5c61"
upstream:
  origin:   "https://github.com/jonathanstowe/p6-Linux-Fuser.git"
  branch:   "master"
  commit:   "98b5c61"
git-subrepo:
  version:  "0.2.3"
  origin:   "git@github.com:ingydotnet/git-subrepo.git"
  commit:   "87a8275"

+++ jonathanstowe/p6-Linux-Fuser/t/01-basic.t
@@ -1,7 +1,7 @@
-use v6;
+use v6.c;
 
 use Test;
-use lib 'lib';
+plan 33;
 
 use Linux::Fuser;
 
@@ -15,50 +15,45 @@ lives-ok { $obj = Linux::Fuser.new() }, "create a new Linux::Fuser";
 
 isa-ok($obj,Linux::Fuser, "and it's the right kind of object");
 
-my $filename = $*PID ~ '.tmp';
+if $*KERNEL.name eq 'linux' {
+    my $filename = $*PID ~ '.tmp';
 
+    my $fh = open $filename, :w;
 
-my $fh = open $filename, :w;
+    do_tests($filename, "with string filename");
+    do_tests($filename.IO, "with IO::Path");
 
-do_tests($filename, "with string filename");
-do_tests($filename.IO, "with IO::Path");
+    sub do_tests(Any $file, Str $description) {
+        my @procs;
+        lives-ok { @procs = $obj.fuser($file) }, "fuser() doesn't die ($description)";
+        ok(@procs.elems, "got some processes ($description)");
+        ok(my $proc = @procs[0], "get the first ($description)");
+        isa-ok($proc,Linux::Fuser::Procinfo, "And it's the right kind of object ($description)");
+        is($proc.pid, $*PID, "got the expected PID ($description)");
+        ok($proc.cmd.elems, "got some command line ($description)");
+        todo("not sure how to test this yet",1);
+        like($proc.cmd[0], /{ $*VM.config<bindir> }/, "and got something like we expected");
+        is($proc.user, $*USER, "got the right user ($description)");
+        ok(my $filedes = $proc.filedes, "get the filedescriptor info ($description)");
+        isa-ok($filedes,Linux::Fuser::FileDescriptor, "and it's the right sort of object ($description)");
+        is($filedes.pos, 0, "pos is 0 ( $description )");
+        ok($filedes.flags > 0, "flags is greater than 0 ($description)");
+        ok($filedes.mnt-id.defined, "mnt_id is defined ( $description)");
+    }
 
-sub do_tests(Any $file, Str $description)
-{
-   my @procs;
+    $fh.close;
+    my @procs;
 
-   lives-ok { @procs = $obj.fuser($file) }, "fuser() doesn't die ($description)";
+    lives-ok { @procs = $obj.fuser($filename) }, "fuser() closed file";
+    is(@procs.elems,0, "and there aren't any processes");
 
-   ok(@procs.elems, "got some processes ($description)");
-   ok(my $proc = @procs[0], "get the first ($description)");
-
-   isa-ok($proc,Linux::Fuser::Procinfo, "And it's the right kind of object ($description)");
-   is($proc.pid, $*PID, "got the expected PID ($description)");
-   ok($proc.cmd.elems, "got some command line ($description)");
-   todo("not sure how to test this yet",1);
-   like($proc.cmd[0], /{ $*VM.config<bindir> }/, "and got something like we expected");
-
-   is($proc.user, $*USER, "got the right user ($description)");
-
-   ok(my $filedes = $proc.filedes, "get the filedescriptor info ($description)");
-   isa-ok($filedes,Linux::Fuser::FileDescriptor, "and it's the right sort of object ($description)");
-   is($filedes.pos, 0, "pos is 0 ( $description )");
-   ok($filedes.flags > 0, "flags is greater than 0 ($description)");
-   ok($filedes.mnt_id.defined, "mnt_id is defined ( $description)");
+    lives-ok { @procs = $obj.fuser('ThiSdoesNotExIst') }, "fuser() no-existent file";
+    is(@procs.elems,0, "and there aren't any processes");
+    $filename.IO.unlink;
 }
-
-$fh.close;
-
-my @procs;
-
-lives-ok { @procs = $obj.fuser($filename) }, "fuser() closed file";
-is(@procs.elems,0, "and there aren't any processes");
-
-lives-ok { @procs = $obj.fuser('ThiSdoesNotExIst') }, "fuser() no-existent file";
-is(@procs.elems,0, "and there aren't any processes");
-
-
-$filename.IO.unlink;
+else {
+    skip-rest "Not linux won't perform tests";
+}
 
 done-testing;
 # vim: expandtab shiftwidth=4 ft=perl6
